change argument order for io.pedestal.app.dataflow/run
The first argument is now the dataflow state. The passed state must be
a map containing the key :data-model instead of just the data-model
map. For example:

In the old version we would pass

{:x 1}

and run would return

{:data-model {:x 2}}

Now we pass

{:data-model {:x 1}}

and run returns

{:data-model {:x 2}}

This allows run to used as a state transition function.