Retain existing service body_params interface while exposing control
This is a bit more work than @r0man's proposal, but it maintains API
compatibility for existing users. Instead of changing existing APIs I
created new custom-{json/edn}-parser functions, with existing functions
being the "default" versions of these parsers. I opted *not* to include
a version for form parsing--we'll cross that bridge when we get to it.

Since it seems to already be an established practice on the project I
preferred key-val pairs over hash-map options, even if at times slightly
more work than necessary.

I also shored up the docs! Thanks again to @r0man for bringing this up
and given me something to do on my plane ride home.

This fixes #96, #97.