Disable service-template tests from running on Travis
Due to some weirdness with service-template tests constantly failing on
Travis (likely on account of technomancy/leiningen#1235) I've decided to
*not* run those tests there and stop hemorrhaging time trying to fix it.

I think this is an acceptable risk to take because service-template is a
very low-churn sub-project. The consequence of this is that project
maintainers will need to be diligent about testing/accepting any pull
requests touching that project. This is similar to the current situation
with app-template (no CI tests.)