Make config.edn -> clj again; Load config when calling (start)
Ran into a problem with edn/read-string and eval. Basically entries like
'project-name.rendering come through edn into ClojureScript as a symbol
with a quote. Attempting to eval the body, however, fails because it
cannot locate the symbols in question

    (eval (edn/read-string "'project-name.rendering"))
    ;; -> CompilerException java.lang.ClassNotFoundException:
    ;;    'project-name.rendering, compiling:(NO_SOURCE_PATH:1:1)

Because of this problem and the annoyance/difficulty of interning
strings as regexps, it seems pertinent to keep the file as Clojure.

Since the config is still executable, but loaded via resources path,
what remains is the question: do we really need the new
expand-app-config functionality, or should we rely solely on functions
called inside that config?