support :read-as option for message parameters in the data renderer
Using this

(msg/param :age {:read-as :data})

will cause the value collected in the Data UI to be read with the
Clojure reader. This allows one to enter arbitrary Clojure in the Data
UI without having to add the conversion code to transform functions.

For now, this should only be used for development.