Fix NPE when path contains trailing slash
Given the following two routes:

- `/people`
- `/people/:name`

It was previously possible to produce an NPE by visiting `/people/`. As
this route is not explicitly defined a 404 response should be returned.

If you wish to use a trailing slash in your route you need to use a
trailing slash in the route definition.