Revert "Support hooking events sent by SSE handlers."
This reverts commit c032b5c4d671c97eaf246bf6318b3f51985babd3.

The implementation of this is ugly, and in any event isn't doing the
right thing. The intention was to support people being able to test
their SSE endpoints. There are two ways to do that: 1) by writing an
"integration" test that interfaces with the target code through the
whole HTTP stack, and 2) by looking at the data that gets put on the
event channel. This change did 2, but in a way that sort of involves the
HTTP-oriented parts of Pedestal. Integration testing is out of scope
here, and the right way to test for case 2 is to put hooks in the
application, not Pedestal.