Handle empty transit body params
When sending a request with a `Content-Type` header of
`application/transit+json` or `application/transit+msgpack` and an empty
body, Pedestal throws a `java.io.EOFException` exception. This happens
because the Transit library doesn't read from an empty
ByteArrayInputStream. This does not happen with EDN, JSON or form
params. This patch fixes that problem by checking the available bytes of
the input stream and makes the behaviour consistent with the other
content types.