Add test for parsing Transit request bodies through response-for.
`response-for` didn't implement `.available` on the ServletInputStream,
which caused it to start failing with the fix to handle empty body
params: d948b6122bbe086562419efa36446f9bd330115c

`response-for` was fixed in 4b5f792dbe81de4c6521aa1e3b82f46eea3e42f6,
and the added test covers this fix. While this latest fix is for
`response-for`, it also provides additional coverage for parsing
(non-empty) Transit bodies.

Change-Id: Ib1fa04f6f52119a9c32a1e2783994f2bacbe306c