Guides should not be misleading
The guides inside the pedestal/pedestal repository are out of date. In
some cases, they mislead users to deprecated features (such as the
`defbefore` helper macro.) In other cases, they simply don't work as
documented.

Since the docs site is now up to date, and we have the
pedestal/pedestal-docs repo, the guides in this repo should be
replaced.

But many blogs and mailing list posts link to the old guides. Therefore,
this patch replaces the text of the guides with links to new docs,
rather than just deleting the old guides.