log-request interceptor made optional.
Problem:
Not every one need the `log-request` interceptor. Since pedestal
is a library it should give the control to user to customize every
aspect of the system ( at least as much as possible ).

Solution:
`log-request` made optional. By default it would be the first interceptor
in the vector returned `default-interceptors`. But if the `::http/no-logging`
key had a truthy value in the service map then log request interceptor
won't be conjed into the default interceptors vector.