Remove deref call in routes function
I may be misunderstanding something but from what I can tell using the `service/routes` symbol directly inside the function will always get the current value of `service/routes` when the `server/routes` function is called, so dereferencing the var is unnecessary. This change works for me - changes to the routes are picked up without a server restart.