Fix unexpanded dev routes
Switching from defroutes to def `[...routes...] means the route is no longer
being expanded at definition time.

This doesn't matter for static routes -- the default interceptor expands them on
startup. But routes wrapped in a function to enable "hot-reloading" aren't, so
we need to do the expansion in the function.