Add clj-kondo linting configuration for the error-dispatch macro
This commit adds a `clj-kondo.exports` folder as explained in `clj-kondo`
documentation. Ideally, if this is bundled into the jar when
library versions are cut, then users will be able to import the config
in their own projects[^1]. This commit shows the example layout of the
config, but more work needs to be done to bundle this into `resources`
available in the final jars.

We use the `:analyze-call` hook[^2] to rewrite the call to
`error-dispatch` as an anonymous function with a `case` statement.
This matches the spirit of the macro-expansion and works for linting
purposes.

Related to: #780

[^1]: Exporting and importing clj-kondo configuration:
https://github.com/clj-kondo/clj-kondo/blob/be6efa075a0a2b903ab220ce7c7e61a242bd9126/doc/config.md#exporting-and-importing-configuration
[^2]: An explanation of the `:analyze-call` hook:
https://github.com/clj-kondo/clj-kondo/blob/be6efa075a0a2b903ab220ce7c7e61a242bd9126/doc/hooks.md#analyze-call